requirements.txt
cfgv==3.4.0
distlib==0.3.7
filelock==3.13.1
identify==2.5.32
nodeenv==1.8.0
platformdirs==4.0.0
pre-commit==3.5.0
PyYAML==6.0.1
virtualenv==20.24.7
pyserial==3.4
